Time: 0 ms (100.00%) | Memory: 8.6 MB (39.50%) - LeetSync

+++ 9-palindrome-number/palindrome-number.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     bool isPalindrome(int x) {
         int dup = x;
-        long long int rev=0;
+        long int rev=0;
         while(x>0){
             int last=x%10;
             rev=(rev*10)+last;
